+++ src/org/eclipse/swt/widgets/Composite.d
@@ -678,7 +678,7 @@ public Control [] getTabList () {
     return tabList;
 }
 
-override int /*long*/ gtk_button_press_event (GtkWidget* widget, GdkEventButton* event) {
+override ptrdiff_t gtk_button_press_event (GtkWidget* widget, GdkEventButton* event) {
     auto result = super.gtk_button_press_event (widget, event);
     if (result !is 0) return result;
     if ((state & CANVAS) !is 0) {
@@ -691,7 +691,7 @@ override int /*long*/ gtk_button_press_event (GtkWidget* widget, GdkEventButton*
     return result;
 }
 
-override int /*long*/ gtk_expose_event (GtkWidget* widget, GdkEventExpose* eventPtr) {
+override ptrdiff_t gtk_expose_event (GtkWidget* widget, GdkEventExpose* eventPtr) {
     if ((state & OBSCURED) !is 0) return 0;
     if ((state & CANVAS) is 0) {
         return super.gtk_expose_event (widget, eventPtr);
@@ -726,7 +726,7 @@ override int /*long*/ gtk_expose_event (GtkWidget* widget, GdkEventExpose* event
     return 0;
 }
 
-override int /*long*/ gtk_key_press_event (GtkWidget* widget, GdkEventKey* event) {
+override ptrdiff_t gtk_key_press_event (GtkWidget* widget, GdkEventKey* event) {
     auto result = super.gtk_key_press_event (widget, event);
     if (result !is 0) return result;
     /*
@@ -747,27 +747,27 @@ override int /*long*/ gtk_key_press_event (GtkWidget* widget, GdkEventKey* event
     return result;
 }
 
-override int /*long*/ gtk_focus (GtkWidget* widget, int directionType) {
+override ptrdiff_t gtk_focus (GtkWidget* widget, int directionType) {
     if (widget is socketHandle) return 0;
     return super.gtk_focus (widget, directionType);
 }
 
-override int /*long*/ gtk_focus_in_event (GtkWidget* widget, GdkEventFocus* event) {
-    int /*long*/ result = super.gtk_focus_in_event (widget, event);
+override ptrdiff_t gtk_focus_in_event (GtkWidget* widget, GdkEventFocus* event) {
+    ptrdiff_t result = super.gtk_focus_in_event (widget, event);
     return (state & CANVAS) !is 0 ? 1 : result;
 }
 
-override int /*long*/ gtk_focus_out_event (GtkWidget* widget, GdkEventFocus* event) {
+override ptrdiff_t gtk_focus_out_event (GtkWidget* widget, GdkEventFocus* event) {
     auto result = super.gtk_focus_out_event (widget, event);
     return (state & CANVAS) !is 0 ? 1 : result;
 }
 
-override int /*long*/ gtk_map (GtkWidget* widget) {
+override ptrdiff_t gtk_map (GtkWidget* widget) {
     fixZOrder ();
     return 0;
 }
 
-override int /*long*/ gtk_realize (GtkWidget* widget) {
+override ptrdiff_t gtk_realize (GtkWidget* widget) {
     auto result = super.gtk_realize (widget);
     if ((style & SWT.NO_BACKGROUND) !is 0) {
         auto window = OS.GTK_WIDGET_WINDOW (paintHandle ());
@@ -779,13 +779,13 @@ override int /*long*/ gtk_realize (GtkWidget* widget) {
     return result;
 }
 
-override int /*long*/ gtk_scroll_child (GtkWidget* widget, int /*long*/ scrollType, int /*long*/ horizontal) {
+override ptrdiff_t gtk_scroll_child (GtkWidget* widget, ptrdiff_t scrollType, ptrdiff_t horizontal) {
     /* Stop GTK scroll child signal for canvas */
     OS.g_signal_stop_emission_by_name (widget, OS.scroll_child.ptr);
     return 1;
 }
 
-override int /*long*/ gtk_style_set (GtkWidget* widget, int /*long*/ previousStyle) {
+override ptrdiff_t gtk_style_set (GtkWidget* widget, ptrdiff_t previousStyle) {
     auto result = super.gtk_style_set (widget, previousStyle);
     if ((style & SWT.NO_BACKGROUND) !is 0) {
         auto window = OS.GTK_WIDGET_WINDOW (paintHandle ());
